\name{CAADate_INT}
\alias{CAADate_INT}
\title{
CAADate_INT
}
\description{
CAADate_INT
}
\usage{
CAADate_INT(value)
}
\arguments{
  \item{value}{
A decimal value
}
}
\details{
}
\value{
the integer part of the decimal number. 
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADate_INT(4.233)
CAADate_INT(4.98)

}
\keyword{ Basic }

