\name{CAADiameters_ApparentAsteroidDiameter}
\alias{CAADiameters_ApparentAsteroidDiameter}
\title{
CAADiameters_ApparentAsteroidDiameter
}
\description{
CAADiameters_ApparentAsteroidDiameter
}
\usage{
CAADiameters_ApparentAsteroidDiameter(Delta, d)
}
\arguments{
  \item{Delta}{
Delta The distance to the asteroid in astronomical units. 
}
  \item{d}{
d The diameter of the asteroid in kilometres.
}
}
\details{
}
\value{
The asteroid's apparent diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_ApparentAsteroidDiameter(Delta = 2, d = 600)
}
\keyword{ Asteroid }

