\name{CAADiameters_AsteroidDiameter}
\alias{CAADiameters_AsteroidDiameter}
\title{
CAADiameters_AsteroidDiameter
}
\description{
CAADiameters_AsteroidDiameter
}
\usage{
CAADiameters_AsteroidDiameter(H, A)
}
\arguments{
  \item{H}{
H The absolute magnitude of the asteroid. 
}
  \item{A}{
A The albedo of the asteroid.
}
}
\details{
}
\value{
The asteroid's diameter in kilometres.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_AsteroidDiameter(H = 5.5, A = 0.6)
}
\keyword{ Asteroid }

