\name{CAADiameters_GeocentricMoonSemidiameter}
\alias{CAADiameters_GeocentricMoonSemidiameter}
\title{
CAADiameters_GeocentricMoonSemidiameter
}
\description{
CAADiameters_GeocentricMoonSemidiameter
}
\usage{
CAADiameters_GeocentricMoonSemidiameter(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to the Moon in kilometres. 
}
}
\details{
}
\value{
The Moon's semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_GeocentricMoonSemidiameter(Delta = 360000)
}
\keyword{ Moon }

