\name{CAADiameters_JupiterEquatorialSemidiameterB}
\alias{CAADiameters_JupiterEquatorialSemidiameterB}
\title{
CAADiameters_JupiterEquatorialSemidiameterB
}
\description{
CAADiameters_JupiterEquatorialSemidiameterB
}
\usage{
CAADiameters_JupiterEquatorialSemidiameterB(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Jupiter in astronomical units. 
}
}
\details{
}
\value{
Jupiter's equatorial semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
Uses the updated Astronomical Almanac for 1984 size.
}
\seealso{

}
\examples{
CAADiameters_JupiterEquatorialSemidiameterB(Delta = 6)
}
\keyword{ Jupiter }

