\name{CAADiameters_JupiterPolarSemidiameterA}
\alias{CAADiameters_JupiterPolarSemidiameterA}
\title{
CAADiameters_JupiterPolarSemidiameterA
}
\description{
CAADiameters_JupiterPolarSemidiameterA
}
\usage{
CAADiameters_JupiterPolarSemidiameterA(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Jupiter in astronomical units. 
}
}
\details{
}
\value{
Jupiter's polar semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_JupiterPolarSemidiameterA(Delta = 6)
}
\keyword{ Jupiter }

