\name{CAADiameters_MarsSemidiameterA}
\alias{CAADiameters_MarsSemidiameterA}
\title{
CAADiameters_MarsSemidiameterA
}
\description{
CAADiameters_MarsSemidiameterA
}
\usage{
CAADiameters_MarsSemidiameterA(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Mars in astronomical units. 
}
}
\details{
}
\value{
Mars's semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_MarsSemidiameterA(Delta = 0.4) 
}
\keyword{ Mars }

