\name{CAADiameters_MarsSemidiameterB}
\alias{CAADiameters_MarsSemidiameterB}
\title{
CAADiameters_MarsSemidiameterB
}
\description{
CAADiameters_MarsSemidiameterB
}
\usage{
CAADiameters_MarsSemidiameterB(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Mars in astronomical units. 
}
}
\details{
}
\value{
Mars's semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
Uses the updated Astronomical Almanac for 1984 size.
}
\seealso{

}
\examples{
CAADiameters_MarsSemidiameterB(Delta = 0.4)
}
\keyword{ Mars }

