\name{CAADiameters_MercurySemidiameterA}
\alias{CAADiameters_MercurySemidiameterA}
\title{
CAADiameters_MercurySemidiameterA
}
\description{
CAADiameters_MercurySemidiameterA
}
\usage{
CAADiameters_MercurySemidiameterA(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Mercury in astronomical units. 
}
}
\details{
}
\value{
Mercury's semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_MercurySemidiameterA(Delta = 0.4)
}
\keyword{ Mercury }

