\name{CAADiameters_MercurySemidiameterB}
\alias{CAADiameters_MercurySemidiameterB}
\title{
CAADiameters_MercurySemidiameterB
}
\description{
CAADiameters_MercurySemidiameterB
}
\usage{
CAADiameters_MercurySemidiameterB(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Mercury in astronomical units. 
}
}
\details{
}
\value{
Mercury's semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
Uses the updated Astronomical Almanac for 1984 size.
}
\seealso{

}
\examples{
CAADiameters_MercurySemidiameterB(Delta = 0.7)
}
\keyword{ Mercury }

