\name{CAADiameters_NeptuneSemidiameterA}
\alias{CAADiameters_NeptuneSemidiameterA}
\title{
CAADiameters_NeptuneSemidiameterA
}
\description{
CAADiameters_NeptuneSemidiameterA
}
\usage{
CAADiameters_NeptuneSemidiameterA(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Neptune in astronomical units. 
}
}
\details{
}
\value{
Neptune's semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_NeptuneSemidiameterA(Delta = 25)
}
\keyword{ Neptune}

