\name{CAADiameters_NeptuneSemidiameterB}
\alias{CAADiameters_NeptuneSemidiameterB}
\title{
CAADiameters_NeptuneSemidiameterB
}
\description{
CAADiameters_NeptuneSemidiameterB
}
\usage{
CAADiameters_NeptuneSemidiameterB(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Neptune in astronomical units.
}
}
\details{
}
\value{
Neptune's semi diameter in arc seconds. 
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
Uses the updated Astronomical Almanac for 1984 size.
}
\seealso{

}
\examples{
CAADiameters_NeptuneSemidiameterB(Delta = 25)
}
\keyword{ Neptune }

