\name{CAADiameters_PlutoSemidiameterB}
\alias{CAADiameters_PlutoSemidiameterB}
\title{
CAADiameters_PlutoSemidiameterB
}
\description{
CAADiameters_PlutoSemidiameterB
}
\usage{
CAADiameters_PlutoSemidiameterB(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Pluto in astronomical units. 
}
}
\details{
}
\value{
Pluto's semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_PlutoSemidiameterB(Delta = 30)
}
\keyword{ Pluto }

