\name{CAADiameters_SaturnEquatorialSemidiameterA}
\alias{CAADiameters_SaturnEquatorialSemidiameterA}
\title{
CAADiameters_SaturnEquatorialSemidiameterA
}
\description{
CAADiameters_SaturnEquatorialSemidiameterA
}
\usage{
CAADiameters_SaturnEquatorialSemidiameterA(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Saturn in astronomical units.  
}
}
\details{
}
\value{
Saturn's equatorial semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_SaturnEquatorialSemidiameterA(Delta = 21)
}
\keyword{ Saturn }

