\name{CAADiameters_SaturnEquatorialSemidiameterB}
\alias{CAADiameters_SaturnEquatorialSemidiameterB}
\title{
CAADiameters_SaturnEquatorialSemidiameterB
}
\description{
CAADiameters_SaturnEquatorialSemidiameterB
}
\usage{
CAADiameters_SaturnEquatorialSemidiameterB(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Saturn in astronomical units. 
}
}
\details{
}
\value{
Saturn's equatorial semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
Uses the updated Astronomical Almanac for 1984 size.
}
\seealso{

}
\examples{
CAADiameters_SaturnEquatorialSemidiameterB(Delta = 21)
}
\keyword{ Saturn }

