\name{CAADiameters_SaturnPolarSemidiameterA}
\alias{CAADiameters_SaturnPolarSemidiameterA}
\title{
CAADiameters_SaturnPolarSemidiameterA
}
\description{
CAADiameters_SaturnPolarSemidiameterA
}
\usage{
CAADiameters_SaturnPolarSemidiameterA(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Saturn in astronomical units.
}
}
\details{
}
\value{
Saturn's polar semi diameter in arc seconds. 
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_SaturnPolarSemidiameterA(Delta = 21)
}
\keyword{ Saturn }

