\name{CAADiameters_SunSemidiameterA}
\alias{CAADiameters_SunSemidiameterA}
\title{
CAADiameters_SunSemidiameterA
}
\description{
CAADiameters_SunSemidiameterA
}
\usage{
CAADiameters_SunSemidiameterA(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to the Sun in astronomical units. 
}
}
\details{
}
\value{
The suns semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_SunSemidiameterA(Delta = 1)
}
\keyword{ Sun }

