\name{CAADiameters_UranusSemidiameterA}
\alias{CAADiameters_UranusSemidiameterA}
\title{
CAADiameters_UranusSemidiameterA
}
\description{
CAADiameters_UranusSemidiameterA
}
\usage{
CAADiameters_UranusSemidiameterA(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Uranus in astronomical units.
}
}
\details{
}
\value{
Uranus's semi diameter in arc seconds .
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_UranusSemidiameterA(Delta = 12)
}
\keyword{ Uranus }

