\name{CAADiameters_UranusSemidiameterB}
\alias{CAADiameters_UranusSemidiameterB}
\title{
CAADiameters_UranusSemidiameterB
}
\description{
CAADiameters_UranusSemidiameterB
}
\usage{
CAADiameters_UranusSemidiameterB(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Uranus in astronomical units.
}
}
\details{
}
\value{
Uranus's semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_UranusSemidiameterB(Delta = 12)
}
\keyword{ Uranus }

