\name{CAADiameters_VenusSemidiameterA}
\alias{CAADiameters_VenusSemidiameterA}
\title{
CAADiameters_VenusSemidiameterA
}
\description{
CAADiameters_VenusSemidiameterA
}
\usage{
CAADiameters_VenusSemidiameterA(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Venus in astronomical units .
}
}
\details{
}
\value{
Venus's semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAADiameters_VenusSemidiameterA(Delta = 0.232)
}
\keyword{ Venus }

