\name{CAADiameters_VenusSemidiameterB}
\alias{CAADiameters_VenusSemidiameterB}
\title{
CAADiameters_VenusSemidiameterB
}
\description{
CAADiameters_VenusSemidiameterB
}
\usage{
CAADiameters_VenusSemidiameterB(Delta)
}
\arguments{
  \item{Delta}{
Delta The distance to Venus in astronomical units .
}
}
\details{
}
\value{
Venus's semi diameter in arc seconds.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
Uses the updated Astronomical Almanac for 1984 size.
}
\seealso{

}
\examples{
CAADiameters_VenusSemidiameterB(Delta = 0.75)
}
\keyword{ Venus }

