\name{CAADynamicalTime_DeltaT}
\alias{CAADynamicalTime_DeltaT}
\title{
Computing DeltaT
}
\description{
Computing Delta T
}
\usage{
CAADynamicalTime_DeltaT(JD)
}

\arguments{
  \item{JD}{
Julian Date
}
}
\details{
see source code of AAplus 1.60.
}
\value{
DeltaT
}
\references{
\url{http://www.naughter.com/aa.html}
}
\author{
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}

\seealso{

}
\examples{
CAADynamicalTime_DeltaT(JD = 2456655.5)
}

\keyword{ time }

