\name{CAADynamicalTime_TAI2TT}
\alias{CAADynamicalTime_TAI2TT}

\title{
TAI to TT
}
\description{
Convert TAI to TT
}
\usage{
CAADynamicalTime_TAI2TT(JD)
}

\arguments{
  \item{JD}{
Julian Date
}
}
\details{
see source code of AAplus 1.60.
}
\value{
TT
}
\references{
    \url{http://www.naughter.com/aa.html}
}
\author{
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}


\seealso{

}
\examples{
CAADynamicalTime_TAI2TT(JD = 2456655.5)
}

\keyword{ time }

