\name{CAADynamicalTime_TT2TAI}
\alias{CAADynamicalTime_TT2TAI}

\title{
Convert TT to TAI
}
\description{
Convert TT to TAI
}
\usage{
CAADynamicalTime_TT2TAI(JD)
}

\arguments{
  \item{JD}{
Julian Date
}
}
\details{
see source code of AAplus 1.60.
}
\value{
TAI
}
\references{
\url{http://www.naughter.com/aa.html}
}
\author{
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}

\seealso{

}
\examples{
CAADynamicalTime_TT2TAI(JD = 2456655.5)
}

\keyword{ time }

