\name{CAADynamicalTime_TT2UT1}
\alias{CAADynamicalTime_TT2UT1}

\title{
TT 2 UT1
}
\description{
TT 2 UT1
}
\usage{
CAADynamicalTime_TT2UT1(JD)
}

\arguments{
  \item{JD}{
Julian Date
}
}
\details{
see source code of AAplus 1.60.
}
\value{
UT1
}
\references{
\url{http://www.naughter.com/aa.html}
}
\author{
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}


\seealso{

}
\examples{
CAADynamicalTime_TT2UT1(JD = 2456655.5)
}

\keyword{ time }

