\name{CAADynamicalTime_TT2UTC}
\alias{CAADynamicalTime_TT2UTC}
\title{
TT to UTC
}
\description{
Convert TT to UTC
}
\usage{
CAADynamicalTime_TT2UTC(JD)
}

\arguments{
  \item{JD}{
Julian Date for TT
}
}
\details{
see the source code of AAplus 1.60
}
\value{
TT
}
\references{
\url{http://www.naughter.com/aa.html}
}
\author{
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{

}


\seealso{
}
\examples{
CAADynamicalTime_TT2UTC(JD = 2456655.5)
}

\keyword{ time }

