\name{CAADynamicalTime_UT12TT}
\alias{CAADynamicalTime_UT12TT}

\title{
UT1 to TT
}
\description{
UT1 to TT
}
\usage{
CAADynamicalTime_UT12TT(JD)
}

\arguments{
  \item{JD}{
Julian Date
}
}
\details{
see source code of AAplus 1.60.
}
\value{
TT
}
\references{
\url{http://www.naughter.com/aa.html}
}
\author{
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}

\seealso{

}
\examples{
CAADynamicalTime_UT12TT(JD = 2456655.5)
}

\keyword{ time }

