\name{CAADynamicalTime_UT1MinusUTC}
\alias{CAADynamicalTime_UT1MinusUTC}

\title{
Difference between UT1 and UTC
}
\description{
UT1 Minus UTC
}
\usage{
CAADynamicalTime_UT1MinusUTC(JD)
}

\arguments{
  \item{JD}{
Julian Date
}
}
\details{
see source code of AAplus 1.60.
}
\value{
UTC
}
\references{
\url{http://www.naughter.com/aa.html}
}
\author{
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}

\seealso{

}
\examples{
CAADynamicalTime_UT1MinusUTC(JD = 2456655.5)

}

\keyword{ time }

