\name{CAADynamicalTime_UTC2TT}
\alias{CAADynamicalTime_UTC2TT}
\title{
UTC to TT
}
\description{
UTC to TT
}
\usage{
CAADynamicalTime_UTC2TT(JD)
}

\arguments{
  \item{JD}{
Julian Date
}
}
\details{
see the source code of AAplus 1.60
}
\value{
JD
}
\references{
\url{http://www.naughter.com/aa.html}
}
\author{
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{

}


\seealso{

}
\examples{
CAADynamicalTime_UTC2TT(JD = 2456655.5)
}

\keyword{ time }

