\name{CAAEarth_Eccentricity}
\alias{CAAEarth_Eccentricity}
\title{
CAAEarth_Eccentricity
}
\description{
CAAEarth_Eccentricity
}
\usage{
CAAEarth_Eccentricity(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for.
JD The date in Dynamical time to calculate for.
}
}
\details{
}
\value{
the eccentricity of Earth's orbit. 
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAEarth_Eccentricity(JD = 2456655.5)
}
\keyword{ Earth }

