\name{CAAEarth_EclipticLongitudeJ2000}
\alias{CAAEarth_EclipticLongitudeJ2000}
\title{
CAAEarth_EclipticLongitudeJ2000
}
\description{
CAAEarth_EclipticLongitudeJ2000
}
\usage{
CAAEarth_EclipticLongitudeJ2000(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for.
JD The date in Dynamical time to calculate for.
}
}
\details{
}
\value{
the ecliptic longitude in degrees referred to the mean dynamical ecliptic and equinox of J2000 in the VSOP theory. 
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAEarth_EclipticLongitudeJ2000(JD = 2456655.5)
}
\keyword{ Earth }

