\name{CAAEarth_RadiusVector}
\alias{CAAEarth_RadiusVector}
\title{
CAAEarth_RadiusVector
}
\description{
CAAEarth_RadiusVector
}
\usage{
CAAEarth_RadiusVector(JD) 
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for.
JD The date in Dynamical time to calculate for.
}
}
\details{
}
\value{
the radius vector in astronomical units.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAEarth_RadiusVector(JD = 2456655.5) 
}
\keyword{ Earth }

