\name{CAAEarth_SunMeanAnomaly}
\alias{CAAEarth_SunMeanAnomaly}
\title{
CAAEarth_SunMeanAnomaly
}
\description{
CAAEarth_SunMeanAnomaly
}
\usage{
CAAEarth_SunMeanAnomaly(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for.
JD The date in Dynamical time to calculate for. 
}
}
\details{
}
\value{
the mean anomaly of the Sun in degrees.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAEarth_SunMeanAnomaly(JD = 2456655.5)
}
\keyword{ Sun }

