\name{CAAElementsPlanetaryOrbit_EarthEccentricity}
\alias{CAAElementsPlanetaryOrbit_EarthEccentricity}
\title{
CAAElementsPlanetaryOrbit_EarthEccentricity
}
\description{
CAAElementsPlanetaryOrbit_EarthEccentricity
}
\usage{
CAAElementsPlanetaryOrbit_EarthEccentricity(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for.
JD The date in Dynamical time to calculate for.
}
}
\details{
}
\value{
the eccentricity of the orbit. 
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_EarthEccentricity(JD = 2456655.5)
}
\keyword{ Earth }

