\name{CAAElementsPlanetaryOrbit_EarthInclination}
\alias{CAAElementsPlanetaryOrbit_EarthInclination}
\title{
CAAElementsPlanetaryOrbit_EarthInclination
}
\description{
CAAElementsPlanetaryOrbit_EarthInclination
}
\usage{
CAAElementsPlanetaryOrbit_EarthInclination()
}
\details{
}
\value{
the inclination on the plane of the ecliptic in degrees with respect to the ecliptic and mean equinox of the date 0.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_EarthInclination()
}
\keyword{ Orbit }

