\name{CAAElementsPlanetaryOrbit_EarthSemimajorAxis}
\alias{CAAElementsPlanetaryOrbit_EarthSemimajorAxis}
\title{
CAAElementsPlanetaryOrbit_EarthSemimajorAxis
}
\description{
CAAElementsPlanetaryOrbit_EarthSemimajorAxis
}
\usage{
CAAElementsPlanetaryOrbit_EarthSemimajorAxis()
}
\details{
}
\value{
the semi major axis of the planet in astronomical units.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_EarthSemimajorAxis()
}
\keyword{ Earth }

