\name{CAAElementsPlanetaryOrbit_JupiterEccentricity}
\alias{CAAElementsPlanetaryOrbit_JupiterEccentricity}
\title{
CAAElementsPlanetaryOrbit_JupiterEccentricity
}
\description{
CAAElementsPlanetaryOrbit_JupiterEccentricity
}
\usage{
CAAElementsPlanetaryOrbit_JupiterEccentricity(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
}
}
\details{
}
\value{
the eccentricity of the orbit.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_JupiterEccentricity(JD = 2456655.5)
}
\keyword{ Jupiter }

