\name{CAAElementsPlanetaryOrbit_JupiterSemimajorAxis}
\alias{CAAElementsPlanetaryOrbit_JupiterSemimajorAxis}
\title{
CAAElementsPlanetaryOrbit_JupiterSemimajorAxis
}
\description{
CAAElementsPlanetaryOrbit_JupiterSemimajorAxis
}
\usage{
CAAElementsPlanetaryOrbit_JupiterSemimajorAxis(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
}
}
\details{
}
\value{
the semi major axis of the planet in astronomical units.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_JupiterSemimajorAxis(JD = 2456655.5)
}
\keyword{ Jupiter }

