\name{CAAElementsPlanetaryOrbit_MercuryInclination}
\alias{CAAElementsPlanetaryOrbit_MercuryInclination}
\title{
CAAElementsPlanetaryOrbit_MercuryInclination
}
\description{
CAAElementsPlanetaryOrbit_MercuryInclination
}
\usage{
CAAElementsPlanetaryOrbit_MercuryInclination(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 

}
}
\details{
}
\value{
the inclination on the plane of the ecliptic in degrees with respect to the ecliptic and mean equinox of the date.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_MercuryInclination(JD = 2456655.5)
}
\keyword{ Mercury }

