\name{CAAElementsPlanetaryOrbit_NeptuneEccentricity}
\alias{CAAElementsPlanetaryOrbit_NeptuneEccentricity}
\title{
CAAElementsPlanetaryOrbit_NeptuneEccentricity
}
\description{
CAAElementsPlanetaryOrbit_NeptuneEccentricity
}
\usage{
CAAElementsPlanetaryOrbit_NeptuneEccentricity(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 

}
}
\details{
}
\value{
the eccentricity of the orbit.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_NeptuneEccentricity(JD = 2456655.5)
}
\keyword{ Neptune }

