\name{CAAElementsPlanetaryOrbit_NeptuneInclinationJ2000}
\alias{CAAElementsPlanetaryOrbit_NeptuneInclinationJ2000}
\title{
CAAElementsPlanetaryOrbit_NeptuneInclinationJ2000
}
\description{
CAAElementsPlanetaryOrbit_NeptuneInclinationJ2000
}
\usage{
CAAElementsPlanetaryOrbit_NeptuneInclinationJ2000(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 

}
}
\details{
}
\value{
the inclination on the plane of the ecliptic in degrees with respect to the fixed ecliptic of J2000.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_NeptuneInclinationJ2000(JD = 2456655.5)
}
\keyword{ Neptune }

