\name{CAAElementsPlanetaryOrbit_NeptuneSemimajorAxis}
\alias{CAAElementsPlanetaryOrbit_NeptuneSemimajorAxis}
\title{
CAAElementsPlanetaryOrbit_NeptuneSemimajorAxis
}
\description{
CAAElementsPlanetaryOrbit_NeptuneSemimajorAxis
}
\usage{ 
CAAElementsPlanetaryOrbit_NeptuneSemimajorAxis(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
JD The date in Dynamical time to calculate for.
}
}
\details{
}
\value{
the semi major axis of the planet in astronomical units.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_NeptuneSemimajorAxis(JD = 2456655.5)
}
\keyword{ Neptune }

