\name{CAAElementsPlanetaryOrbit_SaturnEccentricity}
\alias{CAAElementsPlanetaryOrbit_SaturnEccentricity}
\title{
CAAElementsPlanetaryOrbit_SaturnEccentricity
}
\description{
CAAElementsPlanetaryOrbit_SaturnEccentricity
}
\usage{
CAAElementsPlanetaryOrbit_SaturnEccentricity(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
JD The date in Dynamical time to calculate for.
}
}
\details{
}
\value{
the eccentricity of the orbit.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
} 
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_SaturnEccentricity(JD = 2456655.5)
}
\keyword{ Saturn }

