\name{CAAElementsPlanetaryOrbit_SaturnLongitudeAscendingNode}
\alias{CAAElementsPlanetaryOrbit_SaturnLongitudeAscendingNode}
\title{
CAAElementsPlanetaryOrbit_SaturnLongitudeAscendingNode
}
\description{
CAAElementsPlanetaryOrbit_SaturnLongitudeAscendingNode
}
\usage{
CAAElementsPlanetaryOrbit_SaturnLongitudeAscendingNode(JD)
}
\arguments{ 
  \item{JD}{ The date in Dynamical time to calculate for. 
JD The date in Dynamical time to calculate for.
}
}
\details{
}
\value{
the longitude of the ascending node in degrees with respect to the ecliptic and mean equinox of the date.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_SaturnLongitudeAscendingNode(JD = 2456655.5)
}
\keyword{ Saturn }

