\name{CAAElementsPlanetaryOrbit_SaturnMeanLongitude}
\alias{CAAElementsPlanetaryOrbit_SaturnMeanLongitude}
\title{
CAAElementsPlanetaryOrbit_SaturnMeanLongitude
}
\description{
CAAElementsPlanetaryOrbit_SaturnMeanLongitude
}
\usage{
CAAElementsPlanetaryOrbit_SaturnMeanLongitude(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
JD The date in Dynamical time to calculate for.
}
}
\details{ 
}
\value{
the mean longitude of the planet in degrees with respect to the ecliptic and mean equinox of the date.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_SaturnMeanLongitude(JD = 2456655.5)
}
\keyword{ Saturn }

