\name{CAAElementsPlanetaryOrbit_UranusEccentricity}
\alias{CAAElementsPlanetaryOrbit_UranusEccentricity}
\title{
CAAElementsPlanetaryOrbit_UranusEccentricity
} 
\description{
CAAElementsPlanetaryOrbit_UranusEccentricity
}
\usage{
CAAElementsPlanetaryOrbit_UranusEccentricity(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
JD The date in Dynamical time to calculate for.
}
}
\details{
}
\value{
the eccentricity of the orbit.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_UranusEccentricity(JD = 2456655.5)
}
\keyword{ Uranus }

