\name{CAAElementsPlanetaryOrbit_UranusLongitudePerihelion}
\alias{CAAElementsPlanetaryOrbit_UranusLongitudePerihelion}
\title{
CAAElementsPlanetaryOrbit_UranusLongitudePerihelion
}
\description{
CAAElementsPlanetaryOrbit_UranusLongitudePerihelion
}
\usage{
CAAElementsPlanetaryOrbit_UranusLongitudePerihelion(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
JD The date in Dynamical time to calculate for.
}
}
\details{
}
\value{
the longitude of the perihelion in degrees with respect to the ecliptic and mean equinox of the date. 
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_UranusLongitudePerihelion(JD = 2456655.5)
}
\keyword{ Uranus }

