\name{CAAElementsPlanetaryOrbit_UranusMeanLongitudeJ2000}
\alias{CAAElementsPlanetaryOrbit_UranusMeanLongitudeJ2000}
\title{
CAAElementsPlanetaryOrbit_UranusMeanLongitudeJ2000
}
\description{
CAAElementsPlanetaryOrbit_UranusMeanLongitudeJ2000
}
\usage{
CAAElementsPlanetaryOrbit_UranusMeanLongitudeJ2000(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 

}
}
\details{
}
\value{
the mean longitude of the planet in degrees with respect to the fixed ecliptic of J2000.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_UranusMeanLongitudeJ2000(JD = 2456655.5)
}
\keyword{ Uranus }

