\name{CAAElementsPlanetaryOrbit_VenusLongitudeAscendingNode}
\alias{CAAElementsPlanetaryOrbit_VenusLongitudeAscendingNode}
\title{
CAAElementsPlanetaryOrbit_VenusLongitudeAscendingNode
}
\description{
CAAElementsPlanetaryOrbit_VenusLongitudeAscendingNode
}
\usage{
CAAElementsPlanetaryOrbit_VenusLongitudeAscendingNode(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 

}
}
\details{
}
\value{
the longitude of the ascending node in degrees with respect to the ecliptic and mean equinox of the date.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElementsPlanetaryOrbit_VenusLongitudeAscendingNode(JD = 2456655.5)
}
\keyword{ Venus }

