\name{CAAElementsPlanetaryOrbit_VenusMeanLongitude}
\alias{CAAElementsPlanetaryOrbit_VenusMeanLongitude}
\title{
CAAElementsPlanetaryOrbit_VenusMeanLongitude
}
\description{
CAAElementsPlanetaryOrbit_VenusMeanLongitude
}
\usage{
CAAElementsPlanetaryOrbit_VenusMeanLongitude(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
}
}
\details{
}
\value{
the mean longitude of the planet in degrees with respect to the fixed ecliptic of J2000.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAElementsPlanetaryOrbit_VenusMeanLongitude(JD = 245665.5)
}
\keyword{ Venus }

