\name{CAAElliptical_DistanceToLightTime}
\alias{CAAElliptical_DistanceToLightTime}
\title{
CAAElliptical_DistanceToLightTime
}
\description{
CAAElliptical_DistanceToLightTime
}
\usage{
CAAElliptical_DistanceToLightTime(Distance)
}
\arguments{
  \item{Distance}{
Distance The distance in astronomical units to convert.
}
}
\details{
}
\value{
The light travel time in days corresponding to the distance.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElliptical_DistanceToLightTime(Distance = 6.88)
}
\keyword{ Elliptical }

