\name{CAAElliptical_InstantaneousVelocity}
\alias{CAAElliptical_InstantaneousVelocity}
\title{
CAAElliptical_InstantaneousVelocity
}
\description{
CAAElliptical_InstantaneousVelocity
}
\usage{
CAAElliptical_InstantaneousVelocity(r, a)
}
\arguments{
  \item{r}{
r The distance of the object from the Sun in astronomical units.
}
  \item{a}{
a The semi major axis of the orbit in astronomical units.
}
}
\details{
}
\value{
The instantaneous velocity of the object in kilometres per second.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElliptical_InstantaneousVelocity(r = 3.2, a = 4.4)
}
\keyword{ Elliptical }

