\name{CAAElliptical_LengthOfEllipse}
\alias{CAAElliptical_LengthOfEllipse}
\title{
CAAElliptical_LengthOfEllipse
}
\description{
CAAElliptical_LengthOfEllipse
}
\usage{
CAAElliptical_LengthOfEllipse(e, a)
}
\arguments{
  \item{e}{
e The eccentricity of the orbit.
}
  \item{a}{
a The semi major axis of the orbit in astronomical units.
}
}
\details{
}
\value{
The length of the eclipse in astronomical units.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElliptical_LengthOfEllipse(e  = 0.223, a = 3.2)
}
\keyword{ Elliptical }

