\name{CAAElliptical_MeanMotionFromSemiMajorAxis}
\alias{CAAElliptical_MeanMotionFromSemiMajorAxis}
\title{
CAAElliptical_MeanMotionFromSemiMajorAxis
}
\description{
CAAElliptical_MeanMotionFromSemiMajorAxis
}
\usage{
CAAElliptical_MeanMotionFromSemiMajorAxis(a)
}
\arguments{
  \item{a}{
a The semi major axis of the orbit in astronomical units.
}
}
\details{
}
\value{
The mean motion in degrees / day.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElliptical_MeanMotionFromSemiMajorAxis(a = 2.2)
}
\keyword{ Elliptical }

