\name{CAAElliptical_VelocityAtAphelion}
\alias{CAAElliptical_VelocityAtAphelion}
\title{
CAAElliptical_VelocityAtAphelion
}
\description{
CAAElliptical_VelocityAtAphelion
}
\usage{
CAAElliptical_VelocityAtAphelion(e, a)
}
\arguments{
  \item{e}{
e The eccentricity of the orbit.
}
  \item{a}{
a The semi major axis of the orbit in astronomical units.
}
}
\details{
}
\value{
The velocity of the object in kilometres per second at aphelion.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAElliptical_VelocityAtAphelion(e = 0.22, a = 2.1)
}
\keyword{Aphelion }

