\name{CAAEquinoxesAndSolstices_AutumnEquinox}
\alias{CAAEquinoxesAndSolstices_AutumnEquinox}
\title{
CAAEquinoxesAndSolstices_AutumnEquinox
}
\description{
CAAEquinoxesAndSolstices_AutumnEquinox
}
\usage{
CAAEquinoxesAndSolstices_AutumnEquinox(Year)
}
\arguments{
  \item{Year}{
  Year The year to calculate the occurrence for.
}
}
\details{

}
\value{
The date in Dynamical time when the Autumn Equinox occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAEquinoxesAndSolstices_AutumnEquinox(Year = 2014)
}
\keyword{ Equinox }

