\name{CAAEquinoxesAndSolstices_LengthOfSpring}
\alias{CAAEquinoxesAndSolstices_LengthOfSpring}
\title{
CAAEquinoxesAndSolstices_LengthOfSpring
}
\description{
CAAEquinoxesAndSolstices_LengthOfSpring
}
\usage{
CAAEquinoxesAndSolstices_LengthOfSpring(Year)
}
\arguments{
  \item{Year}{Year The year to calculate for.
}
}
\details{
}
\value{
The length of the astronomical Spring season in days.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAEquinoxesAndSolstices_LengthOfSpring(Year = 2014)
}
\keyword{ Spring }

