\name{CAAEquinoxesAndSolstices_LengthOfSummer}
\alias{CAAEquinoxesAndSolstices_LengthOfSummer}
\title{
CAAEquinoxesAndSolstices_LengthOfSummer
}
\description{
CAAEquinoxesAndSolstices_LengthOfSummer
}
\usage{
CAAEquinoxesAndSolstices_LengthOfSummer(Year)
}
\arguments{
  \item{Year}{ Year The year to calculate for.
}
}
\details{
}
\value{
The length of the astronomical Summer season in days.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAEquinoxesAndSolstices_LengthOfSummer(Year = 2014)
}
\keyword{ Summer }

