\name{CAAEquinoxesAndSolstices_LengthOfWinter}
\alias{CAAEquinoxesAndSolstices_LengthOfWinter}
\title{
CAAEquinoxesAndSolstices_LengthOfWinter
}
\description{
CAAEquinoxesAndSolstices_LengthOfWinter
}
\usage{
CAAEquinoxesAndSolstices_LengthOfWinter(Year)
}
\arguments{
  \item{Year}{ Year The year to calculate for.
}
}
\details{
}
\value{
The length of the astronomical Winter season in days. 
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAEquinoxesAndSolstices_LengthOfWinter(Year = 2014)
}
\keyword{ Winter }

