\name{CAAEquinoxesAndSolstices_NorthernSolstice}
\alias{CAAEquinoxesAndSolstices_NorthernSolstice}

\title{
Northern Solstice
}
\description{
The date in Dynamical time when the sun reaches its highest relative to the celestial equator and appears directly over the Tropic of Cancer.
}
\usage{
CAAEquinoxesAndSolstices_NorthernSolstice(Year)
}

\arguments{
  \item{Year}{
Year The year to calculate the occurrence for. Note that this method refers to "Northern Solstice" instead of "Summer Solstice" to avoid a northern hemisphere-specific bias.
}
}
\details{

}
\value{
The date in Dynamical time when the sun reaches its highest relative to the celestial equator and appears directly over the Tropic of Cancer.
}
\references{
    \url{http://www.naughter.com/aa.html}
}
\author{
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{

}

\seealso{

}
\examples{
 
    CAAEquinoxesAndSolstices_NorthernSolstice(Year = 2000)
 
}

\keyword{ solstice }

