\name{CAAEquinoxesAndSolstices_NorthwardEquinox}
\alias{CAAEquinoxesAndSolstices_NorthwardEquinox}

\title{
Northward Equinox
}
\description{
The date in Dynamical time when the sun crosses the equator in a northerly direction.
}
\usage{
CAAEquinoxesAndSolstices_NorthwardEquinox(Year)
}

\arguments{
  \item{Year}{
Year The year to calculate the occurrence for. Note that this method refers to "Northward Equinox" instead of "Spring Equinox" to avoid a northern hemisphere-specific bias.
}
}
\details{

}
\value{

The date in Dynamical time when the sun crosses the equator in a northerly direction.

}
\references{
    \url{http://www.naughter.com/aa.html}
}
\author{
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{

}

\seealso{
}

\examples{
    CAAEquinoxesAndSolstices_NorthwardEquinox(Year = 2010)
}

\keyword{ NorthwardEquinox }

