\name{CAAEquinoxesAndSolstices_SouthernSolstice}
\alias{CAAEquinoxesAndSolstices_SouthernSolstice}

\title{
Southern Solstice
}
\description{
Southern Solstice
}
\usage{
CAAEquinoxesAndSolstices_SouthernSolstice(Year)
}

\arguments{
  \item{Year}{
    Year The year to calculate the occurrence for. Note that this method refers to "Southern Solstice" instead of "Winter Solstice" to avoid a northern hemisphere-specific bias.

}
}
\details{
to be added.
}
\value{

    The date in Dynamical time when the sun reaches its lowest relative to the celestial equator and appears directly over the Tropic of Capricorn.

}
\references{
    \url{http://www.naughter.com/aa.html}
}
\author{
    C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{

}

\seealso{

}
\examples{

    CAAEquinoxesAndSolstices_SouthernSolstice(Year = 200)

}

\keyword{ Solstice }

