\name{CAAEquinoxesAndSolstices_SouthwardEquinox}
\alias{CAAEquinoxesAndSolstices_SouthwardEquinox}

\title{
Time of Southward Equinox
}
\description{
Time of Southward Equinox
}
\usage{
CAAEquinoxesAndSolstices_SouthwardEquinox(Year)
}

\arguments{
  \item{Year}{
The year to calculate the occurrence for. Note that this method refers to "Southward Equinox" instead of "Fall or Autumn equinox" to avoid a northern hemisphere-specific bias.
}
}
\details{
to be added.
}
\value{
The date in Dynamical time when the Southward Equinox occurs.
}
\references{
    \url{http://www.naughter.com/aa.html}
}
\author{
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{

}


\seealso{

}
\examples{

CAAEquinoxesAndSolstices_SouthwardEquinox(Year = 2000)

}

\keyword{ Equinox }

