\name{CAAEquinoxesAndSolstices_SummerSolstice}
\alias{CAAEquinoxesAndSolstices_SummerSolstice}
\title{
CAAEquinoxesAndSolstices_SummerSolstice
}
\description{
CAAEquinoxesAndSolstices_SummerSolstice
}
\usage{
CAAEquinoxesAndSolstices_SummerSolstice(Year)
}
\arguments{
  \item{Year}{ Year The year to calculate the occurrence for.
}
}
\details{
}
\value{
The date in Dynamical time when the Summer Solstice occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAEquinoxesAndSolstices_SummerSolstice(Year = 2014)
}
\keyword{ Solstice }

