\name{CAAEquinoxesAndSolstices_WinterSolstice}
\alias{CAAEquinoxesAndSolstices_WinterSolstice}
\title{
CAAEquinoxesAndSolstices_WinterSolstice
}
\description{
CAAEquinoxesAndSolstices_WinterSolstice
}
\usage{
CAAEquinoxesAndSolstices_WinterSolstice(Year)
}
\arguments{
  \item{Year}{ 
  Year The year to calculate the occurrence for.
}
}
\details{
}
\value{
The date in Dynamical time when the Winter Solstice occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAEquinoxesAndSolstices_WinterSolstice(Year = 2014)
}
\keyword{ Solstices }

