\name{CAAGlobe_RadiusOfCurvature}
\alias{CAAGlobe_RadiusOfCurvature}
\title{
CAAGlobe_RadiusOfCurvature
}
\description{
CAAGlobe_RadiusOfCurvature
}
\usage{
CAAGlobe_RadiusOfCurvature(GeographicalLatitude)
}
\arguments{
  \item{GeographicalLatitude}{
GeographicalLatitude The latitude of the position in degrees.
}
}
\details{
}
\value{
The radius of curvature of latitude expressed in kilometres.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAGlobe_RadiusOfCurvature(42);
}
\keyword{ Radius }

