\name{CAAGlobe_RadiusOfParallelOfLatitude}
\alias{CAAGlobe_RadiusOfParallelOfLatitude}
\title{
CAAGlobe_RadiusOfParallelOfLatitude
}
\description{
CAAGlobe_RadiusOfParallelOfLatitude
}
\usage{
CAAGlobe_RadiusOfParallelOfLatitude(GeographicalLatitude)
}
\arguments{
  \item{GeographicalLatitude}{
GeographicalLatitude The latitude of the position in degrees.
}
}
\details{
}
\value{
The radius of parallel of latitude expressed in kilometres.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAGlobe_RadiusOfParallelOfLatitude(GeographicalLatitude = 67)
}
\keyword{ latitude }

