\name{CAAIlluminatedFraction_IlluminatedFraction_1}
\alias{CAAIlluminatedFraction_IlluminatedFraction_1}
\title{
CAAIlluminatedFraction_IlluminatedFraction_1
}
\description{
CAAIlluminatedFraction_IlluminatedFraction_1
}
\usage{
CAAIlluminatedFraction_IlluminatedFraction_1(PhaseAngle)
}
\arguments{
  \item{PhaseAngle}{
The phase angle in degrees.
}
}
\details{
}
\value{
The Illuminated Fraction
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAIlluminatedFraction_IlluminatedFraction_1(PhaseAngle = 22)
CAAIlluminatedFraction_IlluminatedFraction_1(PhaseAngle = 180)
CAAIlluminatedFraction_IlluminatedFraction_1(PhaseAngle = 92)
}
\keyword{ Illuminated Fraction }

