\name{CAAIlluminatedFraction_JupiterMagnitudeMuller}
\alias{CAAIlluminatedFraction_JupiterMagnitudeMuller}
\title{
CAAIlluminatedFraction_JupiterMagnitudeMuller
}
\description{
CAAIlluminatedFraction_JupiterMagnitudeMuller
}
\usage{
CAAIlluminatedFraction_JupiterMagnitudeMuller(r, Delta)
}
\arguments{
  \item{r}{
r The planet's distance to the Sun in astronomical units.
}
  \item{Delta}{
Delta The planet's distance from the Earth in astronomical units.
}
}
\details{
}
\value{
The magnitude of the planet.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAIlluminatedFraction_JupiterMagnitudeMuller(r = 16, Delta =14.8 )
}
\keyword{ IlluminatedFraction }
\keyword{ Jupiter }
