\name{CAAIlluminatedFraction_NeptuneMagnitudeAA}
\alias{CAAIlluminatedFraction_NeptuneMagnitudeAA}
\title{
CAAIlluminatedFraction_NeptuneMagnitudeAA
}
\description{
CAAIlluminatedFraction_NeptuneMagnitudeAA
}
\usage{
CAAIlluminatedFraction_NeptuneMagnitudeAA(r, Delta)
}
\arguments{
  \item{r}{
r The planet's distance to the Sun in astronomical units.
}
  \item{Delta}{
Delta The planet's distance from the Earth in astronomical units.
}
}
\details{
}
\value{
The magnitude of the planet.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAIlluminatedFraction_NeptuneMagnitudeAA(r = 23, Delta = 21)
}
\keyword{ IlluminatedFraction }

