\name{CAAIlluminatedFraction_NeptuneMagnitudeMuller}
\alias{CAAIlluminatedFraction_NeptuneMagnitudeMuller}
\title{
CAAIlluminatedFraction_NeptuneMagnitudeMuller
}
\description{
CAAIlluminatedFraction_NeptuneMagnitudeMuller
}
\usage{
CAAIlluminatedFraction_NeptuneMagnitudeMuller(r, Delta)
}
\arguments{
  \item{r}{
r The planet's distance to the Sun in astronomical units.
}
  \item{Delta}{
Delta The planet's distance from the Earth in astronomical units.
}
}
\details{
}
\value{
The magnitude of the planet.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAIlluminatedFraction_NeptuneMagnitudeMuller(r = 22, Delta = 22.5)
}
\keyword{ IlluminatedFraction }

