\name{CAAIlluminatedFraction_PhaseAngle_1}
\alias{CAAIlluminatedFraction_PhaseAngle_1}
\title{
CAAIlluminatedFraction_PhaseAngle_1
}
\description{
CAAIlluminatedFraction_PhaseAngle_1
}
\usage{
CAAIlluminatedFraction_PhaseAngle_1(r, R, Delta)
}
\arguments{
  \item{r}{
r The planet's distance to the Sun in astronomical units.
}
  \item{R}{
R The distance of the Sun from the Earth in astronomical units.
}
  \item{Delta}{
Delta The planet's distance from the Earth in astronomical units.
}
}
\details{
}
\value{
The phase angle in degrees.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated. 
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{

}
\examples{
CAAIlluminatedFraction_PhaseAngle_1(r = 12, R = 1.003, Delta = 12.5)
}
\keyword{ IlluminatedFraction }

