\name{CAAIlluminatedFraction_PlutoMagnitudeAA}
\alias{CAAIlluminatedFraction_PlutoMagnitudeAA}
\title{
CAAIlluminatedFraction_PlutoMagnitudeAA
}
\description{
CAAIlluminatedFraction_PlutoMagnitudeAA
}
\usage{
CAAIlluminatedFraction_PlutoMagnitudeAA(r, Delta)
}
\arguments{
  \item{r}{
r The planet's distance to the Sun in astronomical units.
}
  \item{Delta}{
Delta The planet's distance from the Earth in astronomical units.
}
}
\details{
}
\value{
The magnitude of the planet.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAIlluminatedFraction_PlutoMagnitudeAA(r = 40, Delta = 39)
}
\keyword{ Pluto }

