\name{CAAIlluminatedFraction_UranusMagnitudeMuller}
\alias{CAAIlluminatedFraction_UranusMagnitudeMuller}
\title{
CAAIlluminatedFraction_UranusMagnitudeMuller
}
\description{
CAAIlluminatedFraction_UranusMagnitudeMuller
}
\usage{
CAAIlluminatedFraction_UranusMagnitudeMuller(r, Delta)
}
\arguments{
  \item{r}{
r The planet's distance to the Sun in astronomical units.
}
  \item{Delta}{
Delta The planet's distance from the Earth in astronomical units.
}
}
\details{
}
\value{
The magnitude of the planet.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAIlluminatedFraction_UranusMagnitudeMuller(r = 22.6, Delta = 11)
}
\keyword{ Uranus }

