\name{CAAIlluminatedFraction_VenusMagnitudeAA}
\alias{CAAIlluminatedFraction_VenusMagnitudeAA}
\title{
CAAIlluminatedFraction_VenusMagnitudeAA
}
\description{
CAAIlluminatedFraction_VenusMagnitudeAA
}
\usage{
CAAIlluminatedFraction_VenusMagnitudeAA(r, Delta, i)
}
\arguments{
  \item{r}{
r The planet's distance to the Sun in astronomical units.
}  \item{Delta}{Delta The planet's distance from the Earth in astronomical units.
}  \item{i}{i The planet's phase angle in degrees.}}\details{}\value{The magnitude of the planet.}\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.}\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang}\note{}\seealso{}\examples{CAAIlluminatedFraction_VenusMagnitudeAA(r = 0.75, Delta = 0.43, i = 120)}\keyword{ Venus }