\name{CAAInterpolate_Interpolate5}
\alias{CAAInterpolate_Interpolate5}
\title{
CAAInterpolate_Interpolate5
}
\description{
CAAInterpolate_Interpolate5
}
\usage{
CAAInterpolate_Interpolate5(n, Y1, Y2, Y3, Y4, Y5)
}
\arguments{
  \item{n}{
n The interpolating factor.
}
  \item{Y1}{
Y1 The first Y value to interpolate from.
}
  \item{Y2}{
Y2 The second Y value to interpolate from.
}
  \item{Y3}{
Y3 The third Y value to interpolate from.
}
  \item{Y4}{
Y4 The fourth Y value to interpolate from.
}
  \item{Y5}{
Y5 The fifth Y value to interpolate from.
}
}
\details{
}
\value{
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAInterpolate_Interpolate5(n = 1, Y1 = 9, Y2 = 8, Y3 = 6, Y4 = 2, Y5 = 1)
}
\keyword{ Interpolate }

