\name{CAAInterpolate_Zero2}
\alias{CAAInterpolate_Zero2}
\title{
CAAInterpolate_Zero2
}
\description{
CAAInterpolate_Zero2
}
\usage{
CAAInterpolate_Zero2(Y1, Y2, Y3)
}
\arguments{
  \item{Y1}{
Y1 The first Y value to interpolate from.
}
  \item{Y2}{
Y2 The second Y value to interpolate from.
}
  \item{Y3}{
Y3 The third Y value to interpolate from.
}
}
\details{
}
\value{
The value of the argument X for which the function y becomes zero.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAInterpolate_Zero2(Y1 = 9, Y2 = 3.1, Y3 = -1.9)
}
\keyword{ Interpolate }

