\name{CAAJewishCalendar_DaysInYear}
\alias{CAAJewishCalendar_DaysInYear}
\title{
CAAJewishCalendar_DaysInYear
}
\description{
CAAJewishCalendar_DaysInYear
}
\usage{
CAAJewishCalendar_DaysInYear(Year)
}
\arguments{
  \item{Year}{
 Year The year in the Jewish Calendar.
}
}
\details{
}
\value{
 The number of days in the specified Jewish Year
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAJewishCalendar_DaysInYear(Year = 1309)
}
\keyword{ Calendar }

