\name{CAAJupiter_RadiusVector}
\alias{CAAJupiter_RadiusVector}
\title{
CAAJupiter_RadiusVector
}
\description{
CAAJupiter_RadiusVector
}
\usage{
CAAJupiter_RadiusVector(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 

}
}
\details{
}
\value{
the radius vector in astronomical units.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAJupiter_RadiusVector(JD = 2456597.5)
}
\keyword{ Jupiter }

