\name{CAAMercury_RadiusVector}
\alias{CAAMercury_RadiusVector}
\title{
CAAMercury_RadiusVector
}
\description{
CAAMercury_RadiusVector
}
\usage{
CAAMercury_RadiusVector(JD)
}
\arguments{
  \item{JD}{ 
  JD The date in Dynamical time to calculate for.
}
}
\details{
}
\value{
the radius vector in astronomical units.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMercury_RadiusVector(JD = 2456597.5)
}
\keyword{ Mercury }

