\name{CAAMoon_ArgumentOfLatitude}
\alias{CAAMoon_ArgumentOfLatitude}
\title{
CAAMoon_ArgumentOfLatitude
}
\description{
CAAMoon_ArgumentOfLatitude
}
\usage{
CAAMoon_ArgumentOfLatitude(JD)
}
\arguments{
  \item{JD}{ 
JD The date in Dynamical time to calculate for.
}
}
\details{
}
\value{
the argument of latitude (mean distance of the Moon from its ascending node) in degrees.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoon_ArgumentOfLatitude(JD = 2456597.5)
}
\keyword{ Moon }

