\name{CAAMoon_EclipticLongitude}
\alias{CAAMoon_EclipticLongitude}
\title{
CAAMoon_EclipticLongitude
}
\description{
CAAMoon_EclipticLongitude
}
\usage{
CAAMoon_EclipticLongitude(JD)
}
\arguments{
  \item{JD}{ 
  The date in Dynamical time to calculate for. 
}
}
\details{
}
\value{
the ecliptic longitude in degrees referred to the mean dynamical ecliptic and equinox of the date.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoon_EclipticLongitude(JD = 2456597.5)
}
\keyword{ Moon }

