\name{CAAMoon_MeanAnomaly}
\alias{CAAMoon_MeanAnomaly}
\title{
CAAMoon_MeanAnomaly
}
\description{
CAAMoon_MeanAnomaly
}
\usage{
CAAMoon_MeanAnomaly(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 

}
}
\details{
}
\value{
the mean anomaly of the Moon in degrees.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoon_MeanAnomaly(JD =2456597.5)
}
\keyword{ Moon }

