\name{CAAMoon_MeanElongation}
\alias{CAAMoon_MeanElongation}
\title{
CAAMoon_MeanElongation
}
\description{
CAAMoon_MeanElongation
}
\usage{
CAAMoon_MeanElongation(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 

}
}
\details{
}
\value{
the mean elongation of the Moon in degrees.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoon_MeanElongation(JD = 2456597.5)
}
\keyword{ Moon }

