\name{CAAMoon_MeanLongitudeAscendingNode}
\alias{CAAMoon_MeanLongitudeAscendingNode}
\title{
CAAMoon_MeanLongitudeAscendingNode
}
\description{
CAAMoon_MeanLongitudeAscendingNode
}
\usage{
CAAMoon_MeanLongitudeAscendingNode(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
}
}
\details{
}
\value{
the mean longitude of the ascending node of the Moon in degrees.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoon_MeanLongitudeAscendingNode(JD = 2456597.5)
}
\keyword{ Moon }

