\name{CAAMoon_MeanLongitudePerigee}
\alias{CAAMoon_MeanLongitudePerigee}
\title{
CAAMoon_MeanLongitudePerigee
}
\description{
CAAMoon_MeanLongitudePerigee
}
\usage{
CAAMoon_MeanLongitudePerigee(JD)
}
\arguments{
  \item{JD}{ JD The date in Dynamical time to calculate for.

}
}
\details{
}
\value{
the mean longitude of the perigee of the lunar orbit in degrees. 
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoon_MeanLongitudePerigee(JD  = 2456597.5)
}
\keyword{ Moon }

