\name{CAAMoon_RadiusVectorToHorizontalParallax}
\alias{CAAMoon_RadiusVectorToHorizontalParallax}
\title{
CAAMoon_RadiusVectorToHorizontalParallax
}
\description{
CAAMoon_RadiusVectorToHorizontalParallax
}
\usage{
CAAMoon_RadiusVectorToHorizontalParallax(RadiusVector)
}
\arguments{
  \item{RadiusVector}{
RadiusVector The distance to the object (e.g. Moon) in kilometres.
}
}
\details{
}
\value{
the parallax of the object in degrees.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{ 
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoon_RadiusVectorToHorizontalParallax(RadiusVector = 260077)
}
\keyword{ Moon }

