\name{CAAMoonIlluminatedFraction_IlluminatedFraction}
\alias{CAAMoonIlluminatedFraction_IlluminatedFraction}
\title{
CAAMoonIlluminatedFraction_IlluminatedFraction
}
\description{
CAAMoonIlluminatedFraction_IlluminatedFraction
}
\usage{
CAAMoonIlluminatedFraction_IlluminatedFraction(PhaseAngle)
}
\arguments{
  \item{PhaseAngle}{
PhaseAngle The phase angle in degrees.
}
}
\details{
}
\value{
the illuminated fraction (a value form 0 to 1).
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoonIlluminatedFraction_IlluminatedFraction(PhaseAngle = 123)
}
\keyword{ Moon }

