\name{CAAMoonMaxDeclinations_K}
\alias{CAAMoonMaxDeclinations_K}
\title{
CAAMoonMaxDeclinations_K
}
\description{
CAAMoonMaxDeclinations_K
}
\usage{
CAAMoonMaxDeclinations_K(Year)
}
\arguments{
  \item{Year}{ Year The Year including decimals to calculate the K value for.
}
}
\details{
}
\value{
Returns the approximate value of K (required by the other methods of CAAMoonMaxDeclinations) for calculation of the Moon's max declination.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoonMaxDeclinations_K(Year = 2013)
}
\keyword{ Moon }

