\name{CAAMoonMaxDeclinations_MeanGreatestDeclinationValue}
\alias{CAAMoonMaxDeclinations_MeanGreatestDeclinationValue}
\title{
CAAMoonMaxDeclinations_MeanGreatestDeclinationValue
}
\description{
CAAMoonMaxDeclinations_MeanGreatestDeclinationValue
}
\usage{
CAAMoonMaxDeclinations_MeanGreatestDeclinationValue(k)
}
\arguments{
  \item{k}{
k The K value to calculate the max declination for.
}
}
\details{
}
\value{
Returns the mean maximum declination in degrees.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoonMaxDeclinations_MeanGreatestDeclinationValue(k = 45)
}
\keyword{ Moon }

