\name{CAAMoonNodes_K}
\alias{CAAMoonNodes_K}
\title{
CAAMoonNodes_K
}
\description{
CAAMoonNodes_K
}
\usage{
CAAMoonNodes_K(Year)
}
\arguments{
  \item{Year}{ Year The Year including decimals to calculate the K value for.
}
}
\details{
}
\value{
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoonNodes_K(Year = 2013)
}
\keyword{ Moon }

