\name{CAAMoonNodes_PassageThroNode}
\alias{CAAMoonNodes_PassageThroNode}
\title{
CAAMoonNodes_PassageThroNode
}
\description{
CAAMoonNodes_PassageThroNode
}
\usage{
CAAMoonNodes_PassageThroNode(k)
}
\arguments{
  \item{k}{
k The K value to calculate the passage thro the node for.
}
}
\details{
}
\value{
Returns the date in Dynamical time when the specified passage thro the node occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoonNodes_PassageThroNode(k = 45)
}
\keyword{ Moon }

