\name{CAAMoonPerigeeApogee_ApogeeParallax}
\alias{CAAMoonPerigeeApogee_ApogeeParallax}
\title{
CAAMoonPerigeeApogee_ApogeeParallax
}
\description{
CAAMoonPerigeeApogee_ApogeeParallax
}
\usage{
CAAMoonPerigeeApogee_ApogeeParallax(k)
}
\arguments{
  \item{k}{
k The K value to calculate the apogee parallax for.
}
}
\details{
}
\value{
Returns the parallax at the specified apogee in degrees.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoonPerigeeApogee_ApogeeParallax(k = 45)
}
\keyword{ Moon }

