\name{CAAMoonPerigeeApogee_MeanApogee}
\alias{CAAMoonPerigeeApogee_MeanApogee}
\title{
CAAMoonPerigeeApogee_MeanApogee
}
\description{
CAAMoonPerigeeApogee_MeanApogee
}
\usage{
CAAMoonPerigeeApogee_MeanApogee(k)
}
\arguments{
  \item{k}{
k The K value to calculate the apogee for.
}
}
\details{
}
\value{
Returns the date in Dynamical time when the specified mean apogee occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoonPerigeeApogee_MeanApogee(k = 45)
}
\keyword{ Moon }

