\name{CAAMoonPerigeeApogee_MeanPerigee}
\alias{CAAMoonPerigeeApogee_MeanPerigee}
\title{
CAAMoonPerigeeApogee_MeanPerigee
}
\description{
CAAMoonPerigeeApogee_MeanPerigee
}
\usage{
CAAMoonPerigeeApogee_MeanPerigee(k)
}
\arguments{
  \item{k}{
k The integer value which gives a perigee.
}
}
\details{
}
\value{
Returns the date in Dynamical time when the mean perigee occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoonPerigeeApogee_MeanPerigee(k = 45)
}
\keyword{ Moon }

