\name{CAAMoonPerigeeApogee_PerigeeParallax}
\alias{CAAMoonPerigeeApogee_PerigeeParallax}
\title{
CAAMoonPerigeeApogee_PerigeeParallax
}
\description{
CAAMoonPerigeeApogee_PerigeeParallax
}
\usage{
CAAMoonPerigeeApogee_PerigeeParallax(k)
}
\arguments{
  \item{k}{
k The K value to calculate the perigee parallax for
}
}
\details{
Returns the parallax at the specified perigee in degrees.
}
\value{
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoonPerigeeApogee_PerigeeParallax(k = 45)
}
\keyword{ Moon }

