\name{CAAMoonPerigeeApogee_TruePerigee}
\alias{CAAMoonPerigeeApogee_TruePerigee}
\title{
CAAMoonPerigeeApogee_TruePerigee
}
\description{
CAAMoonPerigeeApogee_TruePerigee
}
\usage{
CAAMoonPerigeeApogee_TruePerigee(k)
}
\arguments{
  \item{k}{
k The K value to calculate the perigee for.
}
}
\details{
}
\value{
Returns the date in Dynamical time when the specified true perigee occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoonPerigeeApogee_TruePerigee(k = 45)
}
\keyword{ Moon }

