\name{CAAMoonPhases_MeanPhase}
\alias{CAAMoonPhases_MeanPhase}
\title{
CAAMoonPhases_MeanPhase
}
\description{
CAAMoonPhases_MeanPhase
}
\usage{
CAAMoonPhases_MeanPhase(k)
}
\arguments{
  \item{k}{
k The K value to calculate the phase for.
}
}
\details{
}
\value{
Returns the date in Dynamical time when the specified mean moon phase occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoonPhases_MeanPhase(k = 45)
}
\keyword{ Moon }

