\name{CAAMoonPhases_TruePhase}
\alias{CAAMoonPhases_TruePhase}
\title{
CAAMoonPhases_TruePhase
}
\description{
CAAMoonPhases_TruePhase
}
\usage{
CAAMoonPhases_TruePhase(k)
}
\arguments{
  \item{k}{
k The K value to calculate the phase for.
}
}
\details{
}
\value{
Returns the date in Dynamical time when the specified true moon phase occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAMoonPhases_TruePhase(k = 5)
}
\keyword{ Moon }

