\name{CAANeptune_RadiusVector}
\alias{CAANeptune_RadiusVector}
\title{
CAANeptune_RadiusVector
}
\description{
CAANeptune_RadiusVector
}
\usage{
CAANeptune_RadiusVector(JD)
}
\arguments{
  \item{JD}{ JD The date in Dynamical time to calculate for.

}
}
\details{
}
\value{
the radius vector in astronomical units.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAANeptune_RadiusVector(JD = 2456597.5)
}
\keyword{ Neptune }

