\name{CAANutation_MeanObliquityOfEcliptic}
\alias{CAANutation_MeanObliquityOfEcliptic}
\title{
CAANutation_MeanObliquityOfEcliptic
}
\description{
CAANutation_MeanObliquityOfEcliptic
}
\usage{
CAANutation_MeanObliquityOfEcliptic(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
}
}
\details{
}
\value{
The mean obliquity of the ecliptic in degrees.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAANutation_MeanObliquityOfEcliptic(JD = 2456597.5)
}
\keyword{ Nutation }

