\name{CAANutation_NutationInLongitude}
\alias{CAANutation_NutationInLongitude}
\title{
CAANutation_NutationInLongitude
}
\description{
CAANutation_NutationInLongitude
}
\usage{
CAANutation_NutationInLongitude(JD)
}
\arguments{
  \item{JD}{ 
  The date in Dynamical time to calculate for. 
}
}
\details{
}
\value{
The nutation in ecliptic longitude in arc seconds of a degree.
}
\references{ 
Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAANutation_NutationInLongitude(JD  = 2456597.5)
}
\keyword{ Nutation }

