\name{CAANutation_NutationInObliquity}
\alias{CAANutation_NutationInObliquity}
\title{
CAANutation_NutationInObliquity
}
\description{
CAANutation_NutationInObliquity
}
\usage{
CAANutation_NutationInObliquity(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
}
}
\details{
}
\value{
The nutation in obliquity in arc seconds of a degree.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAANutation_NutationInObliquity(JD = 2456597.5)
}
\keyword{ Nutation }

