\name{CAANutation_TrueObliquityOfEcliptic}
\alias{CAANutation_TrueObliquityOfEcliptic}
\title{
CAANutation_TrueObliquityOfEcliptic
}
\description{
CAANutation_TrueObliquityOfEcliptic
}
\usage{
CAANutation_TrueObliquityOfEcliptic(JD)
}
\arguments{
  \item{JD}{ 
JD The date in Dynamical time to calculate for.
}
}
\details{
}
\value{
The true obliquity of the ecliptic in degrees.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAANutation_TrueObliquityOfEcliptic(JD = 2456597.5)
}
\keyword{ Nutation }

