\name{CAAParallax_DistanceToParallax}
\alias{CAAParallax_DistanceToParallax}
\title{
CAAParallax_DistanceToParallax
}
\description{
CAAParallax_DistanceToParallax
}
\usage{
CAAParallax_DistanceToParallax(Distance)
}
\arguments{
  \item{Distance}{
Distance The distance (in astronomical units) to the body. 
}
}
\details{
}
\value{
Returns the parallax in degrees. 
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAParallax_DistanceToParallax(Distance = 0.004)
}
\keyword{ Parallax }

