\name{CAAParallax_ParallaxToDistance}
\alias{CAAParallax_ParallaxToDistance}
\title{
CAAParallax_ParallaxToDistance
}
\description{
CAAParallax_ParallaxToDistance
}
\usage{
CAAParallax_ParallaxToDistance(Parallax)
}
\arguments{
  \item{Parallax}{
Parallax The parallax of the body in degrees.
}
}
\details{
}
\value{
Returns the distance in astronomical units.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAParallax_ParallaxToDistance(Parallax = 0.0003)
}
\keyword{ Parallax }

