\name{CAAPhysicalMoon_CalculateGeocentric}
\alias{CAAPhysicalMoon_CalculateGeocentric}
\title{
CAAPhysicalMoon_CalculateGeocentric
}
\description{
CAAPhysicalMoon_CalculateGeocentric
}
\usage{
CAAPhysicalMoon_CalculateGeocentric(JD)
}
\arguments{
  \item{JD}{ 
  The date in Dynamical time to calculate for. 
}
}
\details{
}
\value{
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAPhysicalMoon_CalculateGeocentric(JD = 2456597.5)
}
\keyword{ Moon }

