\name{CAAPhysicalSun_Calculate}
\alias{CAAPhysicalSun_Calculate}
\title{
CAAPhysicalSun_Calculate
}
\description{
CAAPhysicalSun_Calculate
}
\usage{
CAAPhysicalSun_Calculate(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 

}
}
\details{
}
\value{
Time Of Start Of Rotation
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
JD = 2456597.5
CAAPhysicalSun_Calculate(JD)
}
\keyword{ Sun }

