\name{CAAPhysicalSun_TimeOfStartOfRotation}
\alias{CAAPhysicalSun_TimeOfStartOfRotation}
\title{
CAAPhysicalSun_TimeOfStartOfRotation
}
\description{
CAAPhysicalSun_TimeOfStartOfRotation
}
\usage{
CAAPhysicalSun_TimeOfStartOfRotation(C)
}
\arguments{
  \item{C}{
Integer, indicate the Cycles
}
}
\details{
}
\value{
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAPhysicalSun_TimeOfStartOfRotation(C = 67)
}
\keyword{ Rotation }

