\name{CAAPlanetPerihelionAphelion_EarthK}
\alias{CAAPlanetPerihelionAphelion_EarthK}
\title{
CAAPlanetPerihelionAphelion_EarthK
}
\description{
CAAPlanetPerihelionAphelion_EarthK
}
\usage{
CAAPlanetPerihelionAphelion_EarthK(Year)
}
\arguments{
  \item{Year}{ 
  Year The Year including decimals to calculate the K value for.
}
}
\details{
}
\value{
Returns the approximate value of K (required by the other methods of CAAPlanetPerihelionAphelion) for calculation of the dates of Perihelion or Aphelion.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAPlanetPerihelionAphelion_EarthK(Year = 2062)
}
\keyword{ Earth }

