\name{CAAPlanetPerihelionAphelion_JupiterAphelion}
\alias{CAAPlanetPerihelionAphelion_JupiterAphelion}
\title{
CAAPlanetPerihelionAphelion_JupiterAphelion
}
\description{
CAAPlanetPerihelionAphelion_JupiterAphelion
}
\usage{
CAAPlanetPerihelionAphelion_JupiterAphelion(k)
}
\arguments{
  \item{k}{
k The K value to calculate the aphelion for.
}
}
\details{
}
\value{
Returns the date in Dynamical time when the specified aphelion occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{ C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAPlanetPerihelionAphelion_JupiterAphelion(5)
}
\keyword{ Jupiter }
\keyword{ Aphelion }
