\name{CAAPlanetPerihelionAphelion_MercuryAphelion}
\alias{CAAPlanetPerihelionAphelion_MercuryAphelion}
\title{
CAAPlanetPerihelionAphelion_MercuryAphelion
}
\description{
CAAPlanetPerihelionAphelion_MercuryAphelion
}
\usage{
CAAPlanetPerihelionAphelion_MercuryAphelion(k)
}
\arguments{
  \item{k}{
k The K value to calculate the aphelion for.
}
}
\details{
}
\value{
Returns the date in Dynamical time when the specified aphelion occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
options(digits = 15)
CAAPlanetPerihelionAphelion_MercuryAphelion(5)
}
\keyword{ Mercury }
\keyword{ Aphelion }
