\name{CAAPlanetPerihelionAphelion_NeptunePerihelion}
\alias{CAAPlanetPerihelionAphelion_NeptunePerihelion}
\title{
CAAPlanetPerihelionAphelion_NeptunePerihelion
}
\description{
CAAPlanetPerihelionAphelion_NeptunePerihelion
}
\usage{
CAAPlanetPerihelionAphelion_NeptunePerihelion(k)
}
\arguments{
  \item{k}{
k The K value to calculate the aphelion for.
}
}
\details{
}
\value{
Returns the date in Dynamical time when the specified aphelion occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAPlanetPerihelionAphelion_NeptunePerihelion(100)
}
\keyword{ Neptune }
\keyword{ Perihelion }
