\name{CAAPlanetPerihelionAphelion_SaturnAphelion}
\alias{CAAPlanetPerihelionAphelion_SaturnAphelion}
\title{
CAAPlanetPerihelionAphelion_SaturnAphelion
}
\description{
CAAPlanetPerihelionAphelion_SaturnAphelion
}
\usage{
CAAPlanetPerihelionAphelion_SaturnAphelion(k)
}
\arguments{
  \item{k}{
k The K value to calculate the aphelion for.
}
}
\details{
}
\value{
Returns the date in Dynamical time when the specified aphelion occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAPlanetPerihelionAphelion_SaturnAphelion(k = 2)
}
\keyword{ Saturn }
\keyword{ Aphelion }
