\name{CAAPlanetPerihelionAphelion_SaturnPerihelion}
\alias{CAAPlanetPerihelionAphelion_SaturnPerihelion}
\title{
CAAPlanetPerihelionAphelion_SaturnPerihelion
}
\description{
CAAPlanetPerihelionAphelion_SaturnPerihelion
}
\usage{
CAAPlanetPerihelionAphelion_SaturnPerihelion(k)
}
\arguments{
  \item{k}{
k The K value to calculate the perihelion for.
}
}
\details{
}
\value{
Returns the date in Dynamical time when the specified perihelion occurs.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAPlanetPerihelionAphelion_SaturnPerihelion(k = 2)
}
\keyword{ Saturn }
\keyword{ Perihelion }
