\name{CAAPlanetPerihelionAphelion_UranusK}
\alias{CAAPlanetPerihelionAphelion_UranusK}
\title{
CAAPlanetPerihelionAphelion_UranusK
}
\description{
CAAPlanetPerihelionAphelion_UranusK
}
\usage{
CAAPlanetPerihelionAphelion_UranusK(Year)
}
\arguments{
  \item{Year}{ Year The Year including decimals to calculate the K value for.
}
}
\details{
}
\value{
Returns the approximate value of K (required by the other methods of CAAPlanetPerihelionAphelion) for calculation of the dates of Perihelion or Aphelion.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAPlanetPerihelionAphelion_UranusK(Year = 2198)
}
\keyword{ Uranus }

