\name{CAAPluto_EclipticLongitude}
\alias{CAAPluto_EclipticLongitude}
\title{
CAAPluto_EclipticLongitude
}
\description{
CAAPluto_EclipticLongitude
}
\usage{
CAAPluto_EclipticLongitude(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
}
}
\details{
}
\value{
the ecliptic longitude in degrees referred to the standard equinox J2000.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAPluto_EclipticLongitude(JD = 2456597.5)
}
\keyword{ Pluto }

