\name{CAARefraction_RefractionFromTrue}
\alias{CAARefraction_RefractionFromTrue}
\title{
CAARefraction_RefractionFromTrue
}
\description{
CAARefraction_RefractionFromTrue
}
\usage{
CAARefraction_RefractionFromTrue(Altitude, Pressure, Temperature)
}
\arguments{
  \item{Altitude}{
Altitude The true altitude in degrees.
}
  \item{Pressure}{
Pressure The atmospheric pressure in millibars
}
  \item{Temperature}{
Temperature The air temperature in degrees Celsius.
}
}
\details{
}
\value{
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAARefraction_RefractionFromTrue(Altitude = 10, Pressure = 998, Temperature = 12.5)
}
\keyword{ Refraction }

