\name{CAASaturn_RadiusVector}
\alias{CAASaturn_RadiusVector}
\title{
CAASaturn_RadiusVector
}
\description{
CAASaturn_RadiusVector
}
\usage{
CAASaturn_RadiusVector(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
}
}
\details{
}
\value{
the radius vector in astronomical units.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
JD = 2456597.5
CAASaturn_RadiusVector(JD)
}
\keyword{ Saturn }

