\name{CAASATURNPlanet}
\alias{CAASATURNPlanet}
\title{
CAASATURNPlanet
}
\description{
CAASATURNPlanet
}
\usage{
CAASATURNPlanet(Year)
}
\arguments{
  \item{Year}{ Year The year. (Years are counted astronomically i.e. 1 BC = Year 0)
}
}
\details{
}
\value{
A list containing the astronomical phenomena for Saturn.
}
\references{
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAASATURNPlanet(2012)
}
\keyword{ Saturn }
\keyword{ skyevents }
