\name{CAAStellarMagnitudes_BrightnessRatio}
\alias{CAAStellarMagnitudes_BrightnessRatio}
\title{
CAAStellarMagnitudes_BrightnessRatio
}
\description{
CAAStellarMagnitudes_BrightnessRatio
}
\usage{
CAAStellarMagnitudes_BrightnessRatio(m1, m2)
}
\arguments{
  \item{m1}{
m1 The magnitude of the first star.
}
  \item{m2}{
m2  The magnitude of the second star. 
}
}
\details{
}
\value{
the ratio of the apparent luminosities of the two stars.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAStellarMagnitudes_BrightnessRatio(3, 6)
}
\keyword{ Magnitude }

