\name{CAAStellarMagnitudes_CombinedMagnitude}
\alias{CAAStellarMagnitudes_CombinedMagnitude}
\title{
CAAStellarMagnitudes_CombinedMagnitude
}
\description{
CAAStellarMagnitudes_CombinedMagnitude
}
\usage{
CAAStellarMagnitudes_CombinedMagnitude(m1, m2)
}
\arguments{
  \item{m1}{
m1 The magnitude of the first star.
}
  \item{m2}{
m2  The magnitude of the second star.
}
}
\details{
}
\value{ 
the combined magnitude of two stars.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAStellarMagnitudes_CombinedMagnitude(2, 3)
}
\keyword{ Magnitude }

