\name{CAAStellarMagnitudes_MagnitudeDifference}
\alias{CAAStellarMagnitudes_MagnitudeDifference}
\title{
CAAStellarMagnitudes_MagnitudeDifference
}
\description{
CAAStellarMagnitudes_MagnitudeDifference
}
\usage{
CAAStellarMagnitudes_MagnitudeDifference(brightnessRatio)
}
\arguments{
  \item{brightnessRatio}{
brightnessRatio The ratio of the apparent luminosities of the two stars.
}
}
\details{

}
\value{
the difference in magnitude between the stars. 
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAStellarMagnitudes_MagnitudeDifference(brightnessRatio = 2)
}
\keyword{ Magnitude }

