\name{CAAVenus_RadiusVector}
\alias{CAAVenus_RadiusVector}
\title{
CAAVenus_RadiusVector
}
\description{
CAAVenus_RadiusVector
}
\usage{
CAAVenus_RadiusVector(JD)
}
\arguments{
  \item{JD}{ The date in Dynamical time to calculate for. 
}
}
\details{
}
\value{
the radius vector in astronomical units.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
CAAVenus_RadiusVector(2456597.5)
}
\keyword{ Venus }

