\name{JUPITERPlanet}
\alias{JUPITERPlanet}
\title{
JUPITERPlanet
}
\description{
JUPITERPlanet
}
\usage{
JUPITERPlanet(Year)
}
\arguments{
  \item{Year}{ Year The year. (Years are counted astronomically i.e. 1 BC = Year 0)
}
}
\details{
}
\value{
A list containing the astronomical phenomena for Jupiter.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
JUPITERPlanet(2012)
}
\keyword{ Jupiter }
\keyword{ skyevents }
