\name{MERCURYPlanet}
\alias{MERCURYPlanet}
\title{
MERCURYPlanet
}
\description{
MERCURYPlanet
}
\usage{
MERCURYPlanet(Year)
}
\arguments{
  \item{Year}{ Year The year. (Years are counted astronomically i.e. 1 BC = Year 0)
}
}
\details{
}
\value{
A list containing the astronomical phenomena for Mercury.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
MERCURYPlanet(2012)
}
\keyword{ Mercury }
\keyword{ skyevents }
