\name{NEPTUNEPlanet}
\alias{NEPTUNEPlanet}
\title{
NEPTUNEPlanet
}
\description{
NEPTUNEPlanet
}
\usage{
NEPTUNEPlanet(Year)
}
\arguments{
  \item{Year}{ Year The year. (Years are counted astronomically i.e. 1 BC = Year 0)
}
}
\details{
}
\value{
A list containing the astronomical phenomena for Neptune.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
NEPTUNEPlanet(2012)
}
\keyword{ Neptune }
\keyword{ skyevents }
