\name{skycalc-package}
\alias{skycalc-package}
\alias{skycalc}
\docType{package}
\title{
Calculating the Sky Events based on AAplus AA+ v1.56
}
\description{
Implementation of Astronomical Algorithms by Jean Meeus based on AAplus
}
\details{
\tabular{ll}{
Package: \tab skycalc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-1-1\cr
License: \tab GPL-2\cr
}
}
\author{ 
AAplus C++ library by PJ Naughter, imported to R by Jinlong Zhang

Maintainer: Jinlong Zhang <jinlongzhang01@gmail.com>

}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\keyword{ package }
