\name{URANUSPlanet}
\alias{URANUSPlanet}
\title{
URANUSPlanet
}
\description{
URANUSPlanet
}
\usage{
URANUSPlanet(Year)
}
\arguments{
  \item{Year}{ Year The year. (Years are counted astronomically i.e. 1 BC = Year 0)
}
}
\details{
}
\value{
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
URANUSPlanet(1999)
}
\keyword{ Uranus }
\keyword{ skyevents }
