\name{VENUSPlanet}
\alias{VENUSPlanet}
\title{
VENUSPlanet
}
\description{
VENUSPlanet
}
\usage{
VENUSPlanet(Year)
}
\arguments{
  \item{Year}{ Year The year. (Years are counted astronomically i.e. 1 BC = Year 0)
}
}
\details{
}
\value{
A list containing the astronomical phenomena for Venus.
}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\author{  C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{
}
\seealso{
}
\examples{
VENUSPlanet(2012)
}
\keyword{ Venus }
\keyword{ skyevents }
