\name{CAAEquinoxesAndSolstices_LengthOfSpring}
\alias{CAAEquinoxesAndSolstices_LengthOfSpring}

\title{
Length Of Spring
}
\description{
Length Of Spring
}
\usage{
CAAEquinoxesAndSolstices_LengthOfSpring(Year, bNorthernHemisphere = TRUE)
}

\arguments{
  \item{Year}{
The year to calculate for.
}
  \item{bNorthernHemisphere}{
logical, true to indicate the observer is in the northern hemisphere, while false means the southern hemisphere.
}
}
\details{

}
\value{
The length of the astronomical Spring season in days.
}
\references{
    \url{http://www.naughter.com/aa.html}
}
\author{
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{

}

\seealso{
}
\examples{
CAAEquinoxesAndSolstices_LengthOfSpring(Year = 2006, bNorthernHemisphere = TRUE)
}

\keyword{ season }

