\name{CAAEquinoxesAndSolstices_LengthOfWinter}
\alias{CAAEquinoxesAndSolstices_LengthOfWinter}
\title{
Length Of Winter
}
\description{
Length Of Winter
}
\usage{
CAAEquinoxesAndSolstices_LengthOfWinter(Year, bNorthernHemisphere = TRUE)
}

\arguments{
  \item{Year}{
The year to calculate for.
}
  \item{bNorthernHemisphere}{
logical, true to indicate the observer is in the northern hemisphere, while false means the southern hemisphere.
}
}
\details{
to be added.
}
\value{
The length of the astronomical Winter season in days.
}
\references{
    \url{http://www.naughter.com/aa.html}
}
\author{
 C++ code by PJ Naughter, imported to R by Jinlong Zhang
}
\note{

}

\seealso{

}
\examples{

CAAEquinoxesAndSolstices_LengthOfWinter(Year = 2016, bNorthernHemisphere = TRUE)

}

\keyword{ season }

