\name{skycalc-package}
\alias{skycalc-package}
\alias{skycalc}
\docType{package}
\title{
Computational Astronomy in R based on PJ Naughter's AA+ library
}
\description{
skycalc is an R package: a wrapper for the AA+ library for implementing Jean Meeus' Astronomical Algorithms. AA+ was written in C++ by PJ Naughter. 
}
\details{
\tabular{ll}{
Package: \tab skycalc\cr
Type: \tab Package\cr
Version: \tab 1.62\cr
Date: \tab 2015-7-23\cr
License: \tab GPL-2\cr
}
}
\author{ 
AAplus C++ library by PJ Naughter, imported to R by Jinlong Zhang

Maintainer: Jinlong Zhang <jinlongzhang01@gmail.com>

}
\references{ 
 Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.
}
\keyword{ package }
